requirement.txt
click>=8.1.8
google-adk>=1.0.0
google-genai>=1.9.0
python-dotenv>=1.1.0
a2a-sdk
